profile information

+++ student/models.py
@@ -7,7 +7,7 @@ class Student(models.Model):
       user = models.OneToOneField(User, on_delete=models.CASCADE)
       phone = models.CharField(max_length=15)
       country = models.CharField(max_length=50)
-      account_no = models.DecimalField(max_digits=10, decimal_places=2)
+      account_no = models.IntegerField()
       balance = models.DecimalField(max_digits=10, decimal_places=2, default=0)
       image = models.ImageField(upload_to='student/media/images', blank=True)
       slug = models.SlugField(blank=True, null=True, max_length=100)
